[DB] Moved price from REAL to BIGINT to be interpreted as ct, adapted the rest of the code to said change and added possibility of disabling a stock from currently at least being stepped.

+++ api/api.go
@@ -24,7 +24,7 @@ type CreateStockResponse struct {
 }
 type StockGetPriceResponse struct {
 	Code  int
-	Price float64
+	Price int64
 }
 
 type StockGetResponse struct {
